# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "EntitiesPayload": {
        "properties": {
          "entities": {
            "items": {},
            "title": "Entities",
            "type": "array"
          }
        },
        "required": [
          "entities"
        ],
        "title": "EntitiesPayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntentsPayload": {
        "properties": {
          "intents": {
            "items": {},
            "title": "Intents",
            "type": "array"
          }
        },
        "required": [
          "intents"
        ],
        "title": "IntentsPayload",
        "type": "object"
      },
...
      "PatternsPayload": {
        "properties": {
          "patterns": {
            "items": {},
            "title": "Patterns",
            "type": "array"
          }
        },
        "required": [
          "patterns"
        ],
        "title": "PatternsPayload",
        "type": "object"
      },
      "PersonasPayload": {
        "properties": {
          "personas": {
            "items": {},
            "title": "Personas",
            "type": "array"
          }
        },
        "required": [
          "personas"
        ],
        "title": "PersonasPayload",
        "type": "object"
      },
...
      "StatesPayload": {
        "properties": {
          "states": {
            "items": {},
            "title": "States",
            "type": "array"
          }
        },
        "required": [
          "states"
        ],
        "title": "StatesPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "StudioCookie": {
        "in": "cookie",
        "name": "boerdi_studio_auth",
        "type": "apiKey"
      },
      "StudioKey": {
        "in": "header",
        "name": "X-Studio-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "boerdi-chat backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/config/elements": {
      "get": {
        "description": "All editable elements + their source paths for the Studio browser.",
        "operationId": "get_elements_api_config_elements_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Elements Api Config Elements Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Elements",
        "tags": [
          "config-elements"
        ]
      }
    },
    "/api/config/entities": {
      "get": {
        "operationId": "get_entities_route_api_config_entities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Entities Route Api Config Entities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Entities Route",
        "tags": [
          "config-elements"
        ]
      },
      "put": {
        "operationId": "put_entities_route_api_config_entities_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitiesPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Entities Route Api Config Entities Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Put Entities Route",
        "tags": [
          "config-elements"
        ]
      }
    },
...
    "/api/config/intents": {
      "get": {
        "operationId": "get_intents_route_api_config_intents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Intents Route Api Config Intents Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Intents Route",
        "tags": [
          "config-elements"
        ]
      },
      "put": {
        "operationId": "put_intents_route_api_config_intents_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentsPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Intents Route Api Config Intents Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Put Intents Route",
        "tags": [
          "config-elements"
        ]
      }
    },
...
    "/api/config/patterns": {
      "get": {
        "operationId": "get_patterns_route_api_config_patterns_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Patterns Route Api Config Patterns Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Patterns Route",
        "tags": [
          "config-elements"
        ]
      },
      "put": {
        "operationId": "put_patterns_route_api_config_patterns_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatternsPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Patterns Route Api Config Patterns Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Put Patterns Route",
        "tags": [
          "config-elements"
        ]
      }
    },
    "/api/config/personas": {
      "get": {
        "operationId": "get_personas_route_api_config_personas_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Personas Route Api Config Personas Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Personas Route",
        "tags": [
          "config-elements"
        ]
      },
      "put": {
        "operationId": "put_personas_route_api_config_personas_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonasPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Personas Route Api Config Personas Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Put Personas Route",
        "tags": [
          "config-elements"
        ]
      }
    },
...
    "/api/config/states": {
      "get": {
        "operationId": "get_states_route_api_config_states_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get States Route Api Config States Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get States Route",
        "tags": [
          "config-elements"
        ]
      },
      "put": {
        "operationId": "put_states_route_api_config_states_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatesPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put States Route Api Config States Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Put States Route",
        "tags": [
          "config-elements"
        ]
      }
    },
...
  }
}